src/app/components/input-debounce/input-debounce.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { InputDebounceComponent } from './input-debounce.component';

describe('Component: InputDebounce', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [InputDebounceComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([InputDebounceComponent],
      (component: InputDebounceComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(InputDebounceComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(InputDebounceComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-input-debounce></app-input-debounce>
  `,
  directives: [InputDebounceComponent]
})
class InputDebounceComponentTestController {
}

